Adding token requirement for deletions

+++ fso-part04/blog-list/controllers/blogs.js
@@ -32,8 +32,23 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-	await Blog.findByIdAndRemove(request.params.id);
-	response.status(204).end();
+	const blog = await Blog.findById(request.params.id);
+
+	const decodedToken = jwt.verify(request.token, process.env.SECRET);
+	if (!decodedToken.id) {
+		return response.status(401).json({ error: "token invalid" });
+	}
+
+	const user = await User.findById(decodedToken.id);
+
+	if (blog.author.toString() === user.id.toString()) {
+		await Blog.findByIdAndRemove(request.params.id);
+		response.status(204).end();
+	} else {
+		return response.status(401).json({
+			error: "users can only delete their own blogs",
+		});
+	}
 });
 
 blogsRouter.put("/:id", async (request, response) => {
